Alert Service/requirements.txt
quixstreams>=2.3
python-dotenv